requirements.txt
aniso8601==3.0.2
argon2-cffi==19.1.0
cffi==1.12.3
click==6.7
coverage==4.5.1
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-Limiter==1.0.1
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
limits==1.3
MarkupSafe==1.0
peewee==3.7.0
pycparser==2.19
pytz==2018.5
six==1.11.0
Werkzeug==0.14.1
